refactor downloading step
