+++ automation/src/test/scala/org/broadinstitute/dsde/firecloud/page/workspaces/methodconfigs/WorkspaceMethodConfigDetailsPage.scala
@@ -3,11 +3,15 @@ package org.broadinstitute.dsde.firecloud.page.workspaces.methodconfigs
 import java.io.File
 import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter
+import org.broadinstitute.dsde.workbench.service.test.RandomUtil
+
+import org.broadinstitute.dsde.firecloud.fixture.UserFixtures
 
 import com.typesafe.scalalogging.LazyLogging
 import org.broadinstitute.dsde.firecloud.{FireCloudConfig, FireCloudView}
 import org.broadinstitute.dsde.firecloud.component._
 import org.broadinstitute.dsde.firecloud.component.Component._
+import org.broadinstitute.dsde.firecloud.fixture.DownloadFixture
 import org.broadinstitute.dsde.firecloud.page.workspaces.WorkspacePage
 import org.broadinstitute.dsde.firecloud.page.workspaces.monitor.SubmissionDetailsPage
 import org.broadinstitute.dsde.firecloud.page.PageUtil
@@ -20,7 +24,7 @@ import org.scalatest.selenium.Page
 import scala.util.{Failure, Success, Try}
 
 class WorkspaceMethodConfigDetailsPage(namespace: String, name: String, methodConfigNamespace: String, val methodConfigName: String)(implicit webDriver: WebDriver)
-  extends WorkspacePage(namespace, name) with Page with PageUtil[WorkspaceMethodConfigDetailsPage] with LazyLogging with Eventually {
+  extends WorkspacePage(namespace, name) with Page with PageUtil[WorkspaceMethodConfigDetailsPage] with LazyLogging with Eventually with DownloadFixture {
 
   override def awaitReady(): Unit = {
     await condition isLoaded
@@ -158,57 +162,9 @@ class WorkspaceMethodConfigDetailsPage(namespace: String, name: String, methodCo
     modal.importFile(file.getAbsolutePath)
   }
 
-
-  /**
-    * Downloads the metadata currently being viewed.
-    *
-    * If downloadPath is given, the file is given a timestamped name and moved from that location
-    * into the "downloads" directory off the current working directory. This serves two purposes:
-    *
-    * 1. Archiving the file for later inspection when tests fail
-    * 2. Keeping the browser download directory clean so that it doesn't auto-rename subsequent
-    * downloads with the same filename
-    *
-    * @param downloadPath the directory where the browser saves downloaded files
-    * @return the relative path to the moved download file, or None if downloadPath was not given
-    */
-  def downloadInputsJson(downloadPath: Option[String] = None): Option[String] = synchronized {
-
-    def archiveDownloadedFile(sourcePath: String): String = {
-      // wait up to 10 seconds for file exist
-      val f = new File(sourcePath)
-      println("FILE: " + f.getAbsoluteFile)
-      eventually {
-        assert(f.exists(), s"Timed out (10 seconds) waiting for file $f")
-      }
-
-      val date = DateTimeFormatter.ofPattern(dateFormatPatter).format(LocalDateTime.now())
-      val destFile = new File(sourcePath).getName + s".$date"
-      val destPath = s"downloads/$destFile"
-      Util.moveFile(sourcePath, destPath)
-      logger.info(s"Moved file. sourcePath: $sourcePath, destPath: $destPath")
-      destPath
-    }
-
-    downloadInputsJsonLink.awaitEnabled()
-
-    /*
-     * Downloading a file will open another window while the download is in progress and
-     * automatically close it when the download is complete.
-     */
-    // await condition (windowHandles.size == 1, 30)
-    // .submit call takess care waiting for a new window
-    logger.info(s"form: ${form.queryString}")
-    downloadInputsJsonLink.doClick()
-    println("download link click")
-
-    for {
-      path <- downloadPath
-    } yield archiveDownloadedFile(s"$path/inputs.json")
+  def downloadInputsJson(downloadPath: Option[String], fileName: String): Option[String] = {
+    downloadFile(downloadPath, fileName, downloadInputsJsonLink, None)
   }
-
-  lazy val dateFormatPatter = "HH:mm:ss:N" // with nano seconds
-  val form = CssSelectorQuery(s"${inputsTable.query.queryString} form")
 
 
   def changeSnapshotId(newSnapshotId: Int): Unit = {
@@ -223,7 +179,7 @@ class WorkspaceMethodConfigDetailsPage(namespace: String, name: String, methodCo
     }
   }
 
-  // TODO This is a very weak check to deterimine if page is ready
+  // TODO This is a very weak check to determine if page is ready
   def isLoaded: Boolean = {
     await spinner "Loading attributes..."
     openLaunchAnalysisModalButton.isVisible
